+++ src/main/java/com/zestarr/pluginportal/commands/commandutil/SubCommandEnum.java
@@ -1,4 +1,4 @@
-package com.zestarr.pluginportal.commands.commandUtility;
+package com.zestarr.pluginportal.commands.commandutil;
 
 
 import java.util.List;
