+++ src/main/java/com/zestarr/pluginportal/commands/subcommands/InstallSubCommand.java
@@ -1,9 +1,9 @@
-package com.zestarr.pluginportal.commands.ppmSubCommands;
+package com.zestarr.pluginportal.commands.subcommands;
 
 import com.zestarr.pluginportal.PluginPortal;
-import com.zestarr.pluginportal.commands.commandUtility.Flags;
-import com.zestarr.pluginportal.commands.commandUtility.SubCommandEnum;
-import com.zestarr.pluginportal.commands.commandUtility.SubCommandManager;
+import com.zestarr.pluginportal.commands.commandutil.Flags;
+import com.zestarr.pluginportal.commands.commandutil.SubCommandEnum;
+import com.zestarr.pluginportal.commands.commandutil.SubCommandManager;
 import com.zestarr.pluginportal.type.FileType;
 import com.zestarr.pluginportal.type.LocalPlugin;
 import com.zestarr.pluginportal.type.PreviewingPlugin;
