+++ src/main/java/com/zestarr/pluginportal/commands/subcommands/PreviewSubCommand.java
@@ -1,8 +1,8 @@
-package com.zestarr.pluginportal.commands.ppmSubCommands;
+package com.zestarr.pluginportal.commands.subcommands;
 
 import com.zestarr.pluginportal.PluginPortal;
-import com.zestarr.pluginportal.commands.commandUtility.SubCommandEnum;
-import com.zestarr.pluginportal.commands.commandUtility.SubCommandManager;
+import com.zestarr.pluginportal.commands.commandutil.SubCommandEnum;
+import com.zestarr.pluginportal.commands.commandutil.SubCommandManager;
 import com.zestarr.pluginportal.type.PreviewingPlugin;
 import com.zestarr.pluginportal.utils.ChatUtil;
 import org.bukkit.Bukkit;
